requirements.txt
schedule==0.6.0
threadpool==1.3.2
pandas==1.1.0
xlrd==1.2.0
selenium==3.141.0
Pillow==7.2.0
